added documantation to transfer history model
