+++ app/lib/modules/transfer/transfer.dart
@@ -2,3 +2,4 @@ export 'logic/transfer_gistory_store.dart';
 export 'view/transfer_history_view.dart';
 export 'widgets/transaction_card.dart';
 export 'widgets/transactions_empty.dart';
+export 'widgets/transactions_list.dart';
