+++ app/lib/modules/transfer/view/transfer_history_view.dart
@@ -2,7 +2,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
 import 'package:provider/provider.dart';
 
-import 'package:encointer_wallet/models/index.dart';
 import 'package:encointer_wallet/modules/modules.dart';
 import 'package:encointer_wallet/store/app.dart';
 import 'package:encointer_wallet/utils/fetch_status.dart';
@@ -37,24 +36,6 @@ class TransferHistoryView extends StatelessWidget {
             );
         }
       }),
-    );
-  }
-}
-
-class TransactionsList extends StatelessWidget {
-  const TransactionsList({super.key, required this.transactions});
-
-  final List<Transaction> transactions;
-
-  @override
-  Widget build(BuildContext context) {
-    if (transactions.isEmpty) return const TransactionsEmpty();
-    return ListView.builder(
-      itemCount: transactions.length,
-      itemBuilder: (BuildContext context, int index) {
-        final transaction = transactions[index];
-        return TransactionCard(transaction: transaction);
-      },
     );
   }
 }
