+++ app/lib/modules/transfer/widgets/transactions_list.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+
+import 'package:encointer_wallet/models/index.dart';
+import 'package:encointer_wallet/modules/modules.dart';
+
+class TransactionsList extends StatelessWidget {
+  const TransactionsList({super.key, required this.transactions});
+
+  final List<Transaction> transactions;
+
+  @override
+  Widget build(BuildContext context) {
+    if (transactions.isEmpty) return const TransactionsEmpty();
+    return ListView.builder(
+      itemCount: transactions.length,
+      itemBuilder: (BuildContext context, int index) {
+        final transaction = transactions[index];
+        return TransactionCard(transaction: transaction);
+      },
+    );
+  }
+}
